feat: fixed roomNormMapper

+++ department/src/main/kotlin/ru/itmo/department/api/dto/RoomNormResponse.kt
@@ -7,7 +7,7 @@ data class RoomNormResponse(
     var id: UUID? = null,
     var peopleCount: Long? = null,
     var balanceOxygen: Long? = null,
-    var avgPersonNorm: Double? = null,
+    var avgPersonNorm: Long? = null,
     var createdAt: OffsetDateTime? = null,
     var updatedAt: OffsetDateTime? = null
 )
